Move d2l-my-courses-content to main src folder
