+++ src/d2l-my-courses-container.js
@@ -7,7 +7,7 @@ Component for when the `d2l.Tools.MyCoursesWidget.UpdatedSortLogic` config varia
 */
 
 import 'd2l-tabs/d2l-tabs.js';
-import './card-grid/d2l-my-courses-content.js';
+import './d2l-my-courses-content.js';
 import './d2l-utility-behavior.js';
 import './localize-behavior.js';
 import { html, PolymerElement } from '@polymer/polymer/polymer-element.js';
