+++ src/d2l-my-courses-content.js
@@ -11,12 +11,12 @@ import 'd2l-loading-spinner/d2l-loading-spinner.js';
 import 'd2l-simple-overlay/d2l-simple-overlay.js';
 import 'd2l-image-selector/d2l-basic-image-selector.js';
 import 'd2l-typography/d2l-typography-shared-styles.js';
-import '../d2l-all-courses.js';
-import './d2l-card-grid-styles.js';
-import './d2l-card-grid-behavior.js';
-import '../d2l-alert-behavior.js';
-import '../d2l-utility-behavior.js';
-import '../localize-behavior.js';
+import './d2l-all-courses.js';
+import './card-grid/d2l-card-grid-styles.js';
+import './card-grid/d2l-card-grid-behavior.js';
+import './d2l-alert-behavior.js';
+import './d2l-utility-behavior.js';
+import './localize-behavior.js';
 import { Actions } from 'd2l-hypermedia-constants';
 import { html, PolymerElement } from '@polymer/polymer/polymer-element.js';
 import { EnrollmentCollectionEntity } from 'siren-sdk/src/enrollments/EnrollmentCollectionEntity.js';
