+++ test/d2l-my-courses-content/d2l-my-courses-content.html
@@ -5,7 +5,7 @@
 		<script src="../../../@webcomponents/webcomponentsjs/webcomponents-bundle.js"></script>
 		<script src="../../../wct-browser-legacy/browser.js"></script>
 		<script type="module" src="../../../siren-parser/global.js"></script>
-		<script type="module" src="../../src/card-grid/d2l-my-courses-content.js"></script>
+		<script type="module" src="../../src/d2l-my-courses-content.js"></script>
 	</head>
 	<body>
 		<test-fixture id="d2l-my-courses-content-fixture">
